chore: add initial Spring Boot project (#1)
This commit adds the initial Spring Boot project and basic scripts.

The `test.sh` script is not yet implemented, as there are no tests at this time. Ultimately, this script will spin up all dependencies and execute the full test suite. When `test.sh` is added, `localbuild.sh` will be modified to only execute the linter and unit test suite.

Docker operations are handled "manually" at this time to gain experience with Docker. Using the jib-gradle-plugin resulted in a similar sized image. However, it may be beneficial to leverage the plugin in the future.